Extract DrawSelect
